move utils file to general_utils folder
