+++ train_latent_model.py
@@ -11,9 +11,9 @@ import torch.nn.functional as F
 
 from constants import DT
 from constants import PUPPET_GRIPPER_JOINT_OPEN
-from utils import load_data # data functions
-from utils import sample_box_pose, sample_insertion_pose # robot functions
-from utils import compute_dict_mean, set_seed, detach_dict # helper functions
+from general_utils.utils import load_data # data functions
+from general_utils.utils import sample_box_pose, sample_insertion_pose # robot functions
+from general_utils.utils import compute_dict_mean, set_seed, detach_dict # helper functions
 from policy import ACTPolicy, CNNMLPPolicy
 from visualize_episodes import save_videos
 from detr.models.latent_model import Latent_Model_Transformer
